requirements.txt
asgiref==3.7.2
Django==4.2.4
django-environ==0.10.0
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
pytz==2023.3
python-dateutil==2.8.2
tzdata==2023.3
